# apps/web/src/lib/service-clients/service-auth/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "authentication_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/email/generate/link": {
      "post": {
        "tags": ["email::generate_email_link"],
        "summary": "Generates an email link for the user to verify their email address.",
        "operationId": "generate_email_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateEmailLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/jwt/refresh": {
      "post": {
        "tags": ["jwt::refresh"],
        "summary": "Refreshes a JWT token\nYou can either have your access token and refresh token in the cookies or in the request\nheaders\nAuthorization: Bearer <access_token>\nx-macro-refresh-token: <refresh_token>\nThis returns the cookies with the new access and refresh token",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTokensResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Refresh already in progress for this user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/team/invite": {
      "post": {
        "tags": ["teams::inbound::axum_router::invite_to_team"],
        "summary": "Invites a user to a team.",
        "operationId": "invite_to_team",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteToTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "304": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponse": {
        "type": "object",
        "description": "Empty response is required due to custom fetch forcing `response.json()`"
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "A plain old json error response for use with axum.\nyup, thats it.",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message to explain failure"
          }
        }
      },
      "GenerateEmailLinkRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address to resend the verification email to"
          }
        }
      },
...
      "InviteEntry": {
        "type": "object",
        "description": "A single invite entry with email and tier",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user to invite"
          }
        }
      },
      "InviteToTeamRequest": {
        "type": "object",
        "description": "The request body to invite a user to a team",
        "required": ["invites"],
        "properties": {
          "invites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InviteEntry"
            },
            "description": "The invites to send, each with an email and tier"
          }
        }
      },
...
      "UserTokensResponse": {
        "type": "object",
        "required": ["access_token", "refresh_token"],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The users access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "The users refresh token"
          }
        }
      }
    }
  },
...
}